Added all endpoints to axum
