+++ axum-api/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 publish = false
 
 [dependencies]
-axum = "0.6.2"
+axum = { version = "0.6.2", features = ["headers"] }
 axum-macros = "0.3"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.79"
