+++ axum-api/src/main.rs
@@ -1,13 +1,26 @@
-mod add_plan_driver;
+mod drivers;
+mod plans;
+mod schedules;
+mod users;
 
 use axum::{
-    http::StatusCode,
-    response::IntoResponse,
-    routing::{get, post},
-    Extension, Json, Router,
+    async_trait,
+    extract::FromRequestParts,
+    headers::{authorization::Bearer, Authorization},
+    http::{
+        header::{AUTHORIZATION, CONTENT_TYPE},
+        request::Parts,
+        HeaderValue, Method, StatusCode,
+    },
+    routing::{get, patch, post, put},
+    Extension, Router, TypedHeader,
 };
-use serde::{Deserialize, Serialize};
+use data_access::user::Users;
+use endurance_racing_planner_common::{GoogleOpenIdClaims, User};
+use jwt_compact::UntrustedToken;
+use sqlx::PgPool;
 use std::net::SocketAddr;
+use tower_http::cors::CorsLayer;
 
 #[tokio::main]
 async fn main() {
@@ -19,10 +32,22 @@ async fn main() {
     // build our application with a route
     let app = Router::new()
         // `GET /` goes to `root`
-        .route("/", get(root))
-        // `POST /users` goes to `create_user`
-        .route("/users", post(create_user))
-        .route("/plans/:id/drivers", post(add_plan_driver::add_driver))
+        // .route("/", get(root))
+        // // `POST /users` goes to `create_user`
+        // .route("/users", post(create_user))
+        .route("/users/me", get(users::me))
+        .route("/users", post(users::add_user))
+        .route("/plans", get(plans::get_plans))
+        .route("/plans", post(plans::add_plan))
+        .route("/plans/:id", get(plans::get_plan))
+        .route("/plans/:id", patch(plans::patch_plan))
+        .route("/plans/:id/schedule", post(schedules::add_schedule))
+        .route("/plans/:id/schedule", get(schedules::get_schedule))
+        .route("/plans/:id/schedule", put(schedules::put_schedule))
+        .route("/plans/:id/drivers", post(drivers::add_driver))
+        .route("/plans/:id/drivers", get(drivers::get_plan_drivers))
+        .route("/drivers/:id", put(drivers::put_driver))
+        .layer(cors_layer())
         .layer(Extension(db_context));
 
     // run our app with hyper
@@ -35,36 +60,63 @@ async fn main() {
         .unwrap();
 }
 
-// basic handler that responds with a static string
-async fn root() -> &'static str {
-    "Hello, World!"
+fn cors_layer() -> CorsLayer {
+    CorsLayer::new()
+        .allow_headers([CONTENT_TYPE, AUTHORIZATION])
+        .allow_origin("http://localhost:9000".parse::<HeaderValue>().unwrap())
+        .allow_methods([
+            Method::GET,
+            Method::PUT,
+            Method::POST,
+            Method::PATCH,
+            Method::DELETE,
+            Method::OPTIONS,
+        ])
 }
 
-async fn create_user(
-    // this argument tells axum to parse the request body
-    // as JSON into a `CreateUser` type
-    Json(payload): Json<CreateUser>,
-) -> impl IntoResponse {
-    // insert your application logic here
-    let user = User {
-        id: 1337,
-        username: payload.username,
-    };
+pub struct AuthenticatedUser(pub User);
 
-    // this will be converted into a JSON response
-    // with a status code of `201 Created`
-    (StatusCode::CREATED, Json(user))
-}
+#[async_trait]
+impl<B> FromRequestParts<B> for AuthenticatedUser
+where
+    B: Send + Sync,
+{
+    type Rejection = (StatusCode, String);
 
-// the input to our `create_user` handler
-#[derive(Deserialize)]
-struct CreateUser {
-    username: String,
-}
+    async fn from_request_parts(parts: &mut Parts, state: &B) -> Result<Self, Self::Rejection> {
+        let Extension(pool): Extension<PgPool> = Extension::from_request_parts(parts, state)
+            .await
+            .map_err(|err| {
+                tracing::error!(db_error = ?err);
+                (
+                    StatusCode::INTERNAL_SERVER_ERROR,
+                    "database not found".to_string(),
+                )
+            })?;
 
-// the output to our `create_user` handler
-#[derive(Serialize)]
-struct User {
-    id: u64,
-    username: String,
+        let TypedHeader(Authorization(bearer)) =
+            TypedHeader::<Authorization<Bearer>>::from_request_parts(parts, state)
+                .await
+                .map_err(|_| (StatusCode::UNAUTHORIZED, "no bearer token".to_string()))?;
+
+        let oauth_id = UntrustedToken::new(bearer.token())
+            .map_err(|_| (StatusCode::UNAUTHORIZED, "invalid token".to_string()))
+            .and_then(|parsed_token| {
+                parsed_token
+                    .deserialize_claims_unchecked::<GoogleOpenIdClaims>()
+                    .map_err(|_| (StatusCode::UNAUTHORIZED, "invalid token".to_string()))
+            })
+            .and_then(|claims| Ok(claims.custom.sub))?;
+
+        Users::get_user_by_oauth_id(&pool, oauth_id)
+            .await
+            .map_err(|_| {
+                (
+                    StatusCode::INTERNAL_SERVER_ERROR,
+                    "there was a problem locating the user".to_string(),
+                )
+            })
+            .and_then(|user| user.ok_or((StatusCode::UNAUTHORIZED, "user not found".to_string())))
+            .map(Self)
+    }
 }
